# ARROW-11108: [Rust] Fixed performance issue in mutableBuffer.