--- a/rust/arrow/src/array/data.rs
+++ b/rust/arrow/src/array/data.rs
@@ -337,7 +337,7 @@ mod tests {
 
     use std::sync::Arc;
 
-    use crate::datatypes::ToByteSlice;
+    use crate::buffer::Buffer;
     use crate::util::bit_util;
 
     #[test]
@@ -359,7 +359,7 @@ mod tests {
             Some(0),
             None,
             0,
-            vec![Buffer::from([1i32, 2, 3, 4, 5].to_byte_slice())],
+            vec![Buffer::from_slice_ref(&[1i32, 2, 3, 4, 5])],
             vec![],
         ));
         let v = vec![0, 1, 2, 3];