… rust/arrow/src/array/array_union.rs
… rust/arrow/src/array/transform/list.rs
… rust/arrow/src/array/transform/mod.rs
… rust/arrow/src/array/transform/primitive.rs
… rust/arrow/src/array/transform/utils.rs
… rust/arrow/src/array/transform/variable_size.rs
… rust/arrow/src/buffer.rs
… rust/arrow/src/bytes.rs
… rust/arrow/src/compute/kernels/aggregate.rs
… rust/arrow/src/compute/kernels/arithmetic.rs
… rust/arrow/src/compute/kernels/boolean.rs
… rust/arrow/src/compute/kernels/cast.rs
… rust/arrow/src/compute/kernels/comparison.rs
… rust/arrow/src/compute/kernels/filter.rs
… rust/arrow/src/compute/kernels/length.rs
… rust/arrow/src/compute/kernels/limit.rs
… rust/arrow/src/compute/kernels/sort.rs
… rust/arrow/src/compute/kernels/substring.rs
… rust/arrow/src/compute/kernels/take.rs
… rust/arrow/src/compute/util.rs
… rust/arrow/src/datatypes.rs
… rust/arrow/src/json/reader.rs
… rust/arrow/src/lib.rs
… rust/arrow/src/memory.rs
… rust/arrow/src/util/integration_util.rs
… rust/parquet/src/arrow/array_reader.rs
… rust/parquet/src/arrow/record_reader.rs