@Dandandan 2021-01-02
Great! Besides being a speed up, being able to `push`.

I did some quick benchmarking of your branch on DataFusion - I see a speed up of ~5% on q1 / q12.