@Dandandan 2021-01-02
This PR:
```
Query 1 iteration 0 took 608.4 ms
Query 1 iteration 1 took 602.8 ms
Query 1 iteration 2 took 601.4 ms
Query 1 iteration 3 took 604.6 ms
Query 1 iteration 4 took 604.0 ms
Query 1 iteration 5 took 603.6 ms
Query 1 iteration 6 took 605.5 ms
Query 1 iteration 7 took 610.7 ms
Query 1 iteration 8 took 607.5 ms
Query 1 iteration 9 took 607.1 ms
Query 1 avg time: 605.55 ms
```

Master:
```
Query 1 iteration 0 took 642.9 ms
Query 1 iteration 1 took 640.5 ms
Query 1 iteration 2 took 638.1 ms
Query 1 iteration 3 took 639.8 ms
Query 1 iteration 4 took 641.9 ms
Query 1 iteration 5 took 646.4 ms
Query 1 iteration 6 took 638.6 ms
Query 1 iteration 7 took 640.6 ms
Query 1 iteration 8 took 653.0 ms
Query 1 iteration 9 took 648.6 ms
Query 1 avg time: 643.04 ms
```