@jorgecarleitao 2021-01-10
Ok, I have the benchmarks of this PR, which IMO is ready to review. FYI @alamb @nevi-me 

```bash
critcmp master-08cccd68802c9ddc3ca0a5d4bad6e4ba382d74b4 perf_buffer-2dca3c85920c951d92992d343b69a5384178c6fd -t 10
```


```
group                                         master-08cccd68802c9ddc3ca0a5d4bad6e4ba382d74b4    perf_buffer-2dca3c85920c951d92992d343b69a5384178c6fd
-----                                         -----------------------------------------------    ----------------------------------------------------
array_slice 512                               1.17   453.4±37.97ns        ? B/sec                1.00    387.0±8.80ns        ? B/sec
array_string_from_vec 128                     1.28      4.2±0.22µs        ? B/sec                1.00      3.3±0.09µs        ? B/sec
array_string_from_vec 256                     1.45      5.9±0.14µs        ? B/sec                1.00      4.1±0.12µs        ? B/sec
array_string_from_vec 512                     1.66      9.6±0.28µs        ? B/sec                1.00      5.8±0.18µs        ? B/sec
buffer_bit_ops and                            1.23  715.9±278.78ns        ? B/sec                1.00   581.4±16.32ns        ? B/sec
cast date64 to date32 512                     1.11      7.6±0.40µs        ? B/sec                1.00      6.9±0.32µs        ? B/sec
cast float64 to float32 512                   1.21      3.2±0.10µs        ? B/sec                1.00      2.7±0.06µs        ? B/sec
cast float64 to uint64 512                    1.00      3.9±0.41µs        ? B/sec                1.11      4.4±0.12µs        ? B/sec
cast int32 to float32 512                     1.12      3.2±0.11µs        ? B/sec                1.00      2.8±0.08µs        ? B/sec
cast int32 to int64 512                       1.19      3.2±0.08µs        ? B/sec                1.00      2.7±0.07µs        ? B/sec
cast int32 to uint32 512                      1.42      3.9±0.10µs        ? B/sec                1.00      2.8±0.15µs        ? B/sec
cast int64 to int32 512                       1.96      5.1±0.27µs        ? B/sec                1.00      2.6±0.07µs        ? B/sec
cast time32s to time64us 512                  1.11      5.4±0.23µs        ? B/sec                1.00      4.9±0.15µs        ? B/sec
concat i32 1024                               1.00      4.2±0.09µs        ? B/sec                1.12      4.7±0.17µs        ? B/sec
concat str nulls 1024                         1.00     21.3±0.74µs        ? B/sec                1.14     24.3±0.73µs        ? B/sec
eq scalar Float32                             1.26     76.4±5.24µs        ? B/sec                1.00     60.6±3.56µs        ? B/sec
filter context f32 high selectivity           1.00    309.7±1.96µs        ? B/sec                1.12   346.4±10.64µs        ? B/sec
filter context u8 high selectivity            1.27      4.8±0.11µs        ? B/sec                1.00      3.8±0.14µs        ? B/sec
filter context u8 w NULLs high selectivity    1.00    299.2±2.82µs        ? B/sec                1.12    336.0±7.21µs        ? B/sec
filter u8 high selectivity                    1.11     12.2±0.18µs        ? B/sec                1.00     11.0±0.13µs        ? B/sec
from_slice prepared                           1.00   834.5±26.98µs        ? B/sec                1.15   963.4±36.11µs        ? B/sec
gt scalar Float32                             1.13     42.9±1.52µs        ? B/sec                1.00     37.9±1.83µs        ? B/sec
length                                        1.58     57.8±0.58µs        ? B/sec                1.00     36.5±9.01µs        ? B/sec
like_utf8 scalar complex                      1.00  1064.0±91.20µs        ? B/sec                1.17  1241.6±320.92µs        ? B/sec
like_utf8 scalar ends with                    1.14    317.0±8.58µs        ? B/sec                1.00    277.5±5.80µs        ? B/sec
like_utf8 scalar equals                       2.43    142.6±6.62µs        ? B/sec                1.00     58.6±2.81µs        ? B/sec
like_utf8 scalar starts with                  1.41   294.5±16.24µs        ? B/sec                1.00    208.9±7.83µs        ? B/sec
min string 512                                1.16      6.2±0.14µs        ? B/sec                1.00      5.3±0.12µs        ? B/sec
mutable                                       1.19   587.7±15.96µs        ? B/sec                1.00   492.5±17.55µs        ? B/sec
mutable prepared                              1.17   640.6±26.75µs        ? B/sec                1.00   549.0±21.62µs        ? B/sec
mutable str nulls 1024                        1.00      4.0±0.06ms        ? B/sec                1.34      5.3±0.05ms        ? B/sec
neq scalar Float32                            1.23     76.4±3.41µs        ? B/sec                1.00     61.9±6.27µs        ? B/sec
nlike_utf8 scalar starts with                 1.00   321.1±12.95µs        ? B/sec                1.27    407.8±9.19µs        ? B/sec
struct_array_from_vec 1024                    1.53     20.8±0.55µs        ? B/sec                1.00     13.6±0.66µs        ? B/sec
struct_array_from_vec 128                     1.16      7.2±0.21µs        ? B/sec                1.00      6.2±0.23µs        ? B/sec
struct_array_from_vec 256                     1.27      9.1±0.29µs        ? B/sec                1.00      7.1±0.20µs        ? B/sec
struct_array_from_vec 512                     1.40     13.0±0.27µs        ? B/sec                1.00      9.3±0.32µs        ? B/sec
take i32 1024                                 1.26      2.4±0.96µs        ? B/sec                1.00  1883.9±35.60ns        ? B/sec
```