@alamb 2021-01-11
FWIW it might be good to wait to merge this PR until after 3.0 is shipped in case it introduces any subtle new bugs we would have longer to shake the out